Moved interface into session.
